Split into multiple files, refactor
